:heavy_plus_sign: Add Reports class
Also, fix update to reflect the modern API which only exports one report at a time
Closes #168
