+++ tests/unit/test_long_project.py
@@ -124,6 +124,6 @@ def test_import_dataframe(long_project):
 
 
 def test_reports_df_export(long_project):
-    report = long_project.export_reports(report_id="1", format="df")
+    report = long_project.export_report(report_id="1", format="df")
 
     assert isinstance(report, pd.DataFrame)
